Remove Clutter from device-manager, first working version
