+++ ImageManager/ErrorManager.py
