+++ ImageManager/__init__.py
