+++ ImageManager/conf.py
