+++ ImageManager/utils.py
